Added Sweet alert Changes

+++ src/app/components/contacts/edit-contact/edit-contact.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Contact } from 'src/app/models/contacts.model';
 import { ContactsService } from 'src/app/services/contacts.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-edit-contact',
@@ -39,24 +40,71 @@ export class EditContactComponent implements OnInit {
   }
 
   updateContact() {
-    this.contactsService.updateContact(this.contactDetail.id, this.contactDetail).subscribe({
-      next: (contact) => {
-        this.router.navigate(["contacts"]);
-      },
-      error: (response) => {
-        console.log(response);
-      }
-    })
+    if(this.validateContact()){
+      this.contactsService.updateContact(this.contactDetail.id, this.contactDetail).subscribe({
+        next: (contact) => {
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'Contact Updated Successfully',
+            showConfirmButton: false,
+            timer: 1500,
+            customClass: 'swal-wide',
+          })
+          this.router.navigate(["contacts"]);
+        },
+        error: (response) => {
+          console.log(response);
+        }
+      })
+    }
+    else{
+      Swal.fire({
+        icon: 'error',
+        title: 'Please Fill Needed Details',
+        showConfirmButton: true,
+        customClass: 'swal-wide',
+        confirmButtonColor:"#0D6EFD",
+      })
+    }
   }
 
   deleteContact(id:string) {
-    this.contactsService.deleteContact(id).subscribe({
-      next: (contact) => {
-        this.router.navigate(["contacts"]);
-      },
-      error: (response) => {
-        console.log(response);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      customClass: 'swal-wide',
+      confirmButtonColor:"#0D6EFD",
+      cancelButtonColor:"#DC3545",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.contactsService.deleteContact(id).subscribe({
+          next: (contact) => {
+            Swal.fire({
+              position: 'top-end',
+              icon: 'success',
+              title: 'Contact Deleted Successfully',
+              showConfirmButton: false,
+              timer: 1500,
+              customClass: 'swal-wide',
+            })
+            this.router.navigate(["contacts"]);
+          },
+          error: (response) => {
+            console.log(response);
+          }
+        })
       }
     })
+    
+  }
+  validateContact():boolean{
+    if(this.contactDetail.firstName.trim() != "" && this.contactDetail.lastName.trim() != "" && this.contactDetail.email.trim() != "" && this.contactDetail.phoneNumber.trim() != ""){
+      return true;
+    }
+    return false;
   }
 }
